requirements.txt
python-telegram-bot
fastapi
uvicorn
motor
python-dotenv
pydantic
httpx
gunicorn